refactor: remove output port directory for token
